Separates parse_input into chunks
Also cleans up imports
